pyproject.toml
[project]
name = "dechaos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "google-adk",
    "litellm>=1.76.0",
    "pdfplumber",
]
